+++ docs/changes.md
@@ -3,6 +3,7 @@
 2023/12/13
   * 增加csv row object，方便在MVVM中创建对象。
   * 增加函数csv\_file\_object\_find\_first
+  * fscript 增加函数sha256。
 
 2023/12/11
   * 完善 cvs file object 保存命令。